fix(hydra): gate statfs disk-pressure code to unix builds
The statfs disk-pressure code added in
https://github.com/helixml/helix/pull/2712 calls syscall.Statfs and
syscall.Statfs_t, both of which are unix-only in the Go stdlib. The
helix CLI release-backend pipeline cross-compiles for windows/amd64,
and hydra is pulled into the CLI via the import graph, so the windows
build fails:

  Building helix-windows-amd64.exe
  # github.com/helixml/helix/api/pkg/hydra
  api/pkg/hydra/disk_pressure.go:222: undefined: syscall.Statfs_t

Tag build 2467 for 2.11.25 failed for this reason.

Fix: split the statfs functions out of disk_pressure.go into two
build-tagged files. disk_pressure_statfs_unix.go (//go:build unix)
owns the real implementations. disk_pressure_statfs_windows.go
(//go:build windows) provides stubs that return a "not supported on
windows" error so the dispatcher falls open. Hydra never actually runs
on windows; the CLI just has to compile.

Verified clean cross-compile across all five release targets:
linux/amd64, linux/arm64, darwin/amd64, darwin/arm64, windows/amd64.
Hydra unit tests pass unchanged.
